Hide tooltip and create label-with-values option
BSS adds a title to the button it generates. Some browsers will turn this into a browser generated tooltip. I added the ability to turn this off while maintaining backwards compatibility. Added another data attribute for label-with-values that is specific to Power search, These are quick fixes and can be expanded on later.